box: make xrow_update_init and xrow_update_read_ops non-static
Also fix compilation of rope.h as C++, because now it is included into:
 1. src/box/xrow_update_field.h:88
 2. src/box/xrow_update.h:37
 3. src/box/tuple.h:1548
 4. src/box/tuple_compare.cc:33

Needed for tarantool/tarantool-ee#628

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring
